% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsglm.R
\name{print.CCFFbinsglm}
\alias{print.CCFFbinsglm}
\title{Internal function.}
\usage{
\method{print}{CCFFbinsglm}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinsglm} objects.}
}
\description{
Internal function.
}
\keyword{internal}
