% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binspwc.R
\name{print.CCFFbinspwc}
\alias{print.CCFFbinspwc}
\title{Internal function.}
\usage{
\method{print}{CCFFbinspwc}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinspwc} objects.}
}
\description{
Internal function.
}
\keyword{internal}
