% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsqreg.R
\name{print.CCFFbinsqreg}
\alias{print.CCFFbinsqreg}
\title{Internal function.}
\usage{
\method{print}{CCFFbinsqreg}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinsqreg} objects.}
}
\description{
Internal function.
}
\keyword{internal}
