% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsreg.R
\name{print.CCFFbinsreg}
\alias{print.CCFFbinsreg}
\title{Internal function.}
\usage{
\method{print}{CCFFbinsreg}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinsreg} objects.}
}
\description{
Internal function.
}
\keyword{internal}
