% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsregselect.R
\name{print.CCFFbinsregselect}
\alias{print.CCFFbinsregselect}
\title{Internal function.}
\usage{
\method{print}{CCFFbinsregselect}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinsregselect} objects.}
}
\description{
Internal function.
}
\keyword{internal}
