% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsregtest.R
\name{print.CCFFbinsregtest}
\alias{print.CCFFbinsregtest}
\title{Internal function.}
\usage{
\method{print}{CCFFbinsregtest}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinsregtest} objects.}
}
\description{
Internal function.
}
\keyword{internal}
