% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binstest.R
\name{print.CCFFbinstest}
\alias{print.CCFFbinstest}
\title{Internal function.}
\usage{
\method{print}{CCFFbinstest}(x, ...)
}
\arguments{
\item{x}{Class \code{CCFFbinstest} objects.}
}
\description{
Internal function.
}
\keyword{internal}
