% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsglm.R
\name{summary.CCFFbinsglm}
\alias{summary.CCFFbinsglm}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinsglm}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinsglm} objects.}
}
\description{
Internal function.
}
\keyword{internal}
