% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binspwc.R
\name{summary.CCFFbinspwc}
\alias{summary.CCFFbinspwc}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinspwc}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinspwc} objects.}
}
\description{
Internal function.
}
\keyword{internal}
