% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsqreg.R
\name{summary.CCFFbinsqreg}
\alias{summary.CCFFbinsqreg}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinsqreg}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinsqreg} objects.}
}
\description{
Internal function.
}
\keyword{internal}
