% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsreg.R
\name{summary.CCFFbinsreg}
\alias{summary.CCFFbinsreg}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinsreg}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinsreg} objects.}
}
\description{
Internal function.
}
\keyword{internal}
