% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsregselect.R
\name{summary.CCFFbinsregselect}
\alias{summary.CCFFbinsregselect}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinsregselect}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinsregselect} objects.}
}
\description{
Internal function.
}
\keyword{internal}
