% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsregtest.R
\name{summary.CCFFbinsregtest}
\alias{summary.CCFFbinsregtest}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinsregtest}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinsregtest} objects.}
}
\description{
Internal function.
}
\keyword{internal}
