% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binstest.R
\name{summary.CCFFbinstest}
\alias{summary.CCFFbinstest}
\title{Internal function.}
\usage{
\method{summary}{CCFFbinstest}(object, ...)
}
\arguments{
\item{object}{Class \code{CCFFbinstest} objects.}
}
\description{
Internal function.
}
\keyword{internal}
